+++ Anoushka/Project_flask/templates/login.html
@@ -21,35 +21,12 @@
                         <div class="invalid-feedback">
                             
                         </div>
-                        <!-- <div class="input-group" >
-                            <div class="input-icon">
-                                <span class="input-group-text"><i class='bx bxs-user'></i></span>
-                            </div>
-                        </div> -->
                     <input type="'text" class="input" name="username" placeholder="Username" required="required" value="{{request.form.username}}">
                     <span class="input-group-text"><i class='bx bxs-user'></i></span>
                 </div>
                 <div class="form-group">
-                    <!-- 
-                        <div class="input-group">
-                            <div class="input-icon">
-                                <span class="input-group-text"><i class='bx bxs-lock-open-alt'></i></span>
-                            </div>
-                        </div>
-                        <div class="invalid-feedback">
-                            
-                        </div>
-                        <div class="input-group">
-                                <span class="input-group-text"><i class='bx bxs-lock-open-alt'></i></span>
-                            </div>
-                        </div>-->
                     <input type="password" class="input" name="password" placeholder="Password" required="required" value="{{request.form.password}}">
                     <span class="input-group-text"><i class='bx bxs-lock-open-alt'></i></span>
-                </div>
-                <div class="form-check">
-                   <div id="redirect">
-                        <a href="{{ url_for('reset_request') }}">Forgot Password?</a>
-                    </div>
                 </div>
             </fieldset>
             <div class="btn">
