+++ Anoushka/flask_original_py_files/watchharbour_deleted/models_deleted.py
@@ -1,10 +1,13 @@
+# from itsdangerous import TimedJSONWebSignatureSerializer as Serializer
 from Project_flask import db, login_manager
 from flask_login import UserMixin
+
 
 
 @login_manager.user_loader
 def load_user(user_id):
     return User.query.get(int(user_id))
+
 
 class User(db.Model, UserMixin):
     id = db.Column(db.Integer, primary_key=True)
@@ -12,6 +15,19 @@ class User(db.Model, UserMixin):
     email = db.Column(db.String(120), unique=True, nullable=False)
     image_file = db.Column(db.String(20), nullable=False, default='default.jpeg')
     password = db.Column(db.String(60), nullable=False)
+
+    # def get_reset_token(self, expires_sec=1800):
+    #     # s = Serializer(app.config['SECRET_KEY'], expires_sec)
+    #     return s.dumps({'user_id': self.id}).decode('utf-8')
+
+    # @staticmethod
+    # def verify_reset_token(token):
+    #     # s = Serializer(app.config['SECRET_KEY'])
+    #     try:
+    #         user_id = s.loads(token)['user_id']
+    #     except:
+    #         return None
+    #     return User.query.get(user_id)
 
     def __repr__(self):
         return f"User('{self.username}', '{self.email}', '{self.image_file}')"
